server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "uvicorn>=0.32.1",
    "emails>=0.6.0",
    "jinja2>=3.1.2",
    "bcrypt>=4.3.0",
    "reportlab>=4.4.3",
    "python-dotenv>=1.1.0",
]
